fix: adding string key to unclassified model (NLU-321)
