+++ packages/base-types/src/models/version/nluUnclassifiedData.ts
@@ -1,0 +1,19 @@
+export enum NLUUnclassifiedDataType {
+  CONVERSATION = 'conversation',
+  PROTOTYPE = 'prototype',
+  NLU_DATASOURCE_IMPORT = 'nluDatasourceImport',
+}
+
+export interface NLUUnclassifiedUtterances {
+  utterance: string;
+  sourceID?: string;
+}
+
+export interface NLUUnclassifiedData {
+  key?: string;
+  creatorID?: number;
+  type: NLUUnclassifiedDataType;
+  name: string;
+  utterances: NLUUnclassifiedUtterances[];
+  importedAt?: string;
+}
